Added body parsing. Needs more testing.
There is also a bug in the current body parsing logic.
The parser right now parses the body till the length is equal to the content-length.
But if the body is bigger it should return an error.

This happens not always. If a packet is received and the packet reaches
the requested in the content-length even though there would still be packets send afterwards,
the body is parsed and no error is returned.

This needs to be fixed but needs some refactoring.